We have added an inventorycontainer interface to indicate that these classes contain an inventory. They should also implement the "getInventory" method from the interface to pass the inventory to whichever object wants it.
What this allows us to do is treat players, stores, and any other class we create in the future the same way. It allows us to implement a service that lets inventorycontainers trade with each other. And if at any point we want to implement something new, like a creature, we can give them the inventorycontainer interface, and we can give them items, and let them trade with stores or players as we see fit.